fixed rubocop, improved modularization... =)
